이윤서: (엔진)충돌 구조 업데이트, 마우스 커서 렌더링

+++ GameEngineContents/GlobalActor.cpp
@@ -24,7 +24,7 @@ bool GlobalActor::PixelCollisionMapUpdate(GlobalActor* _Actor)
 		MsgBoxAssert("충돌맵이 설정되지 않았습니다");
 	}
 
-	float4 Color = MapTexture_->GetPixel(dynamic_cast<GlobalActor*>(_Actor)->GetTransform().GetWorldPosition().ix(), -(dynamic_cast<GlobalActor*>(_Actor)->GetTransform().GetWorldPosition().iy()) + 39.f);
+	float4 Color = MapTexture_->GetPixel((float)dynamic_cast<GlobalActor*>(_Actor)->GetTransform().GetWorldPosition().ix(), -(float)(dynamic_cast<GlobalActor*>(_Actor)->GetTransform().GetWorldPosition().iy()) + 39.f);
 
 	if (false == Color.CompareInt4D(float4{1.f, 1.f, 1.f, 0.f}))
 	{
